chore: move the backend food crate into a virtual workspace
Note that sqlx is not the best at working in a workspace, to make tests
that requires migrations, I need to have the migrations in the workspace
root for the build, I think, but also in the backend crate directory for
integration tests, which is sub-optimal for sure but easy to solve using
a symbolic link. Sorry not sorry for everyone on Windows where symbolic
links are still a second class citizen.